Run f (#24)
* Update princess.py with Run F

Updated run f with the forge alignment. Changed a few angles, checked the angle accuracy, and adjusted a few distances.

* Update princess.py

Updated run f code to go back to base and to complete ore pick up.

* Update princess.py

Made changes in order to adjust the angle when aligning to forge.

* Update Run F with new arrangment

Rearranged the code in order to do forge first to ensure that our alignment to forge is accurate.

* Update princess.py 

Started to work on hitting the Silo lever.

* Refactor run_f function to code silo and heavy lifting

Still working on completing the ore drop off, but completed silo, heavy lifting, who lived here, and releasing the ore.

* Refactor hammer movement and motor parameters

Comment out hammer movement commands and adjust parameters for motor functions.

---------

Co-authored-by: raajv07 <rajagopal.vijayaraghavan@gmail.com>